+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/ExportClient/OtlpHttpMetricsExportClient.cs
@@ -1,4 +1,4 @@
-// <copyright file="OtlpHttpTraceExportClient.cs" company="OpenTelemetry Authors">
+// <copyright file="OtlpHttpMetricsExportClient.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -25,25 +25,25 @@ using System.Threading;
 #endif
 using System.Threading.Tasks;
 using Google.Protobuf;
-using OtlpCollector = Opentelemetry.Proto.Collector.Trace.V1;
+using OtlpCollector = Opentelemetry.Proto.Collector.Metrics.V1;
 
 namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClient
 {
-    /// <summary>Class for sending OTLP trace export request over HTTP.</summary>
-    internal sealed class OtlpHttpTraceExportClient : BaseOtlpHttpExportClient<OtlpCollector.ExportTraceServiceRequest>
+    /// <summary>Class for sending OTLP metrics export request over HTTP.</summary>
+    internal sealed class OtlpHttpMetricsExportClient : BaseOtlpHttpExportClient<OtlpCollector.ExportMetricsServiceRequest>
     {
         internal const string MediaContentType = "application/x-protobuf";
-        private readonly Uri exportTracesUri;
+        private readonly Uri exportMetricsUri;
 
-        public OtlpHttpTraceExportClient(OtlpExporterOptions options, HttpClient httpClient)
+        public OtlpHttpMetricsExportClient(OtlpExporterOptions options, HttpClient httpClient)
             : base(options, httpClient)
         {
-            this.exportTracesUri = this.Options.Endpoint;
+            this.exportMetricsUri = this.Options.Endpoint;
         }
 
-        protected override HttpRequestMessage CreateHttpRequest(OtlpCollector.ExportTraceServiceRequest exportRequest)
+        protected override HttpRequestMessage CreateHttpRequest(OtlpCollector.ExportMetricsServiceRequest exportRequest)
         {
-            var request = new HttpRequestMessage(HttpMethod.Post, this.exportTracesUri);
+            var request = new HttpRequestMessage(HttpMethod.Post, this.exportMetricsUri);
             foreach (var header in this.Headers)
             {
                 request.Headers.Add(header.Key, header.Value);
@@ -58,9 +58,9 @@ namespace OpenTelemetry.Exporter.OpenTelemetryProtocol.Implementation.ExportClie
         {
             private static readonly MediaTypeHeaderValue ProtobufMediaTypeHeader = new MediaTypeHeaderValue(MediaContentType);
 
-            private readonly OtlpCollector.ExportTraceServiceRequest exportRequest;
+            private readonly OtlpCollector.ExportMetricsServiceRequest exportRequest;
 
-            public ExportRequestContent(OtlpCollector.ExportTraceServiceRequest exportRequest)
+            public ExportRequestContent(OtlpCollector.ExportMetricsServiceRequest exportRequest)
             {
                 this.exportRequest = exportRequest;
                 this.Headers.ContentType = ProtobufMediaTypeHeader;