+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.2.0-rc1
+
+Released 2021-Nov-29
+
 * Added configuration options for `MetricReaderType` to allow for configuring
   the `ConsoleMetricExporter` to export either manually or periodically.
   ([#2648](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2648))